# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  auth.JWTResponse:
    properties:
      access_token:
        type: string
    type: object
  auth.LoginAccesoRequest:
    properties:
      acceso_id:
        type: integer
      clave_kiosko:
        type: string
    required:
    - acceso_id
    - clave_kiosko
    type: object
  auth.LoginRequest:
    properties:
      correo:
        type: string
      password:
        type: string
    required:
    - correo
    - password
    type: object
  auth.RegisterRequest:
    properties:
      correo:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - correo
    - password
    type: object
  auth.SesionResponse:
    properties:
      token:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend del proyecto AUTOnomIA para la FEPRO 2026 de la BUAP. Gestiona
    el registro y control de visitantes/accesos en condominios.
  title: AUTOnomIA API
  version: "1.0"
paths:
...
  /auth/acceso/{id}/revocar:
    post:
      description: Revoca todas las sesiones activas del Acceso indicado, para un
        kiosko robado o perdido
      parameters:
      - description: ID del acceso
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Revoca las sesiones de un kiosko
      tags:
      - auth
  /auth/acceso/login:
    post:
      consumes:
      - application/json
      description: Da acceso a un kiosko a partir del ID de su Acceso y su clave,
        abriendo una sesion revocable
      parameters:
      - description: AccesoID y ClaveKiosko
        in: body
        name: acceso
        required: true
        schema:
          $ref: '#/definitions/auth.LoginAccesoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.SesionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Loguea a un kiosko
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Da acceso a un admin a partir de su correo y contraseña, devolviendo
        un JWT
      parameters:
      - description: Correo y password del Admin
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.JWTResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Loguea a un admin
      tags:
      - auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Registra un nuevo administrador, hasheando su password con bcrypt,
        y devuelve su JWT
      parameters:
      - description: Correo y password del nuevo Admin
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.JWTResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Crear admin
      tags:
      - auth
...
swagger: "2.0"